requirements.txt
requests
anthropic
resend
python-dotenv
feedparser
